Fix bug regarding string logging twice
The cause of the bug is my placement of the logs inside the playRound function. It is called twice inside the game function hence running to the log inside the playRound function twice for draw and loss outcomes. Moving the logs to the game function results in the correct output and cleaner code.